+++ src/main/java/br/com/zalf/prolog/webservice/frota/veiculo/VeiculoResource.java
@@ -82,11 +82,11 @@ public final class VeiculoResource {
             Pilares.Frota.Veiculo.ALTERAR,
             Pilares.Frota.Veiculo.CADASTRAR,
             Pilares.Frota.Checklist.VISUALIZAR_TODOS})
-    @Path("/{codUnidade}")
-    public List<Veiculo> getVeiculosAtivosByUnidade(@HeaderParam("Authorization") @Required String userToken,
-                                                    @PathParam("codUnidade") @Required Long codUnidade,
-                                                    @QueryParam("ativos") @Optional Boolean ativos) {
-        return service.getVeiculosAtivosByUnidade(userToken, codUnidade, ativos);
+    @Path("/busca/byUnidade")
+    public List<VeiculoListagem> buscaVeiculosByUnidade(@HeaderParam("Authorization") @Required String userToken,
+                                                        @QueryParam("codUnidade") @Required Long codUnidade,
+                                                        @QueryParam("somenteAtivos") @Optional Boolean somenteAtivos) {
+        return service.buscaVeiculosByUnidade(userToken, codUnidade, somenteAtivos);
     }
 
     @GET
@@ -187,15 +187,15 @@ public final class VeiculoResource {
             Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_VEICULO_ESTOQUE,
             Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_ANALISE,
             Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_DESCARTE})
-    @Path("/com-pneus/{placa}")
+    @Path("/visualizacao")
     @AppVersionCodeHandler(
             implementation = DefaultAppVersionCodeHandler.class,
             targetVersionCode = 68,
             versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
             actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
-    public Veiculo getVeiculoByPlacaComPneus(@HeaderParam("Authorization") String userToken,
-                                             @PathParam("placa") String placa) {
-        return service.getVeiculoByPlaca(userToken, placa, true);
+    public VeiculoVisualizacao buscaVeiculoByCodigo(@HeaderParam("Authorization") String userToken,
+                                                        @QueryParam("codVeiculo") Long codVeiculo) {
+        return service.buscaVeiculoByCodigo(userToken, codVeiculo);
     }
 
     @GET
@@ -206,6 +206,31 @@ public final class VeiculoResource {
     public Veiculo getVeiculoByPlacaSemPneus(@HeaderParam("Authorization") String userToken,
                                              @PathParam("placa") String placa) {
         return service.getVeiculoByPlaca(userToken, placa, false);
+    }
+
+    /**
+     * @deprecated at 2020-05-07.
+     * <p>
+     * Este método foi depreciado pois um novo foi criado: {@link #buscaVeiculoByCodigo(String, Long)}
+     */
+    @Deprecated
+    @GET
+    @Secured(permissions = {Pilares.Frota.Veiculo.VISUALIZAR,
+            Pilares.Frota.Veiculo.CADASTRAR,
+            Pilares.Frota.Veiculo.ALTERAR,
+            Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_VEICULO_ESTOQUE,
+            Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_VEICULO_ESTOQUE,
+            Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_ANALISE,
+            Pilares.Frota.Pneu.Movimentacao.MOVIMENTAR_DESCARTE})
+    @Path("/com-pneus/{placa}")
+    @AppVersionCodeHandler(
+            implementation = DefaultAppVersionCodeHandler.class,
+            targetVersionCode = 68,
+            versionCodeHandlerMode = VersionCodeHandlerMode.BLOCK_THIS_VERSION_AND_BELOW,
+            actionIfVersionNotPresent = VersionNotPresentAction.BLOCK_ANYWAY)
+    public Veiculo getVeiculoByPlacaComPneus(@HeaderParam("Authorization") String userToken,
+                                             @PathParam("placa") String placa) {
+        return service.getVeiculoByPlaca(userToken, placa, true);
     }
 
     /**
@@ -228,5 +253,24 @@ public final class VeiculoResource {
     @Path("/marcaModelos/{codEmpresa}")
     public List<Marca> getMarcaModeloVeiculoByCodEmpresa(@PathParam("codEmpresa") Long codEmpresa) {
         return service.getMarcaModeloVeiculoByCodEmpresa(codEmpresa);
+    }
+
+    /**
+     * @deprecated at 2020-05-07.
+     * <p>
+     * Este método foi depreciado pois um novo foi criado: {@link #buscaVeiculosAtivosByUnidade(String, Long, Boolean)}
+     */
+    @Deprecated
+    @GET
+    @Secured(permissions = {
+            Pilares.Frota.Veiculo.VISUALIZAR,
+            Pilares.Frota.Veiculo.ALTERAR,
+            Pilares.Frota.Veiculo.CADASTRAR,
+            Pilares.Frota.Checklist.VISUALIZAR_TODOS})
+    @Path("/{codUnidade}")
+    public List<Veiculo> getVeiculosAtivosByUnidade(@HeaderParam("Authorization") @Required String userToken,
+                                                    @PathParam("codUnidade") @Required Long codUnidade,
+                                                    @QueryParam("ativos") @Optional Boolean ativos) {
+        return service.getVeiculosAtivosByUnidade(userToken, codUnidade, ativos);
     }
 }